+++ R/condPlotCat2.R
@@ -189,5 +189,3 @@ unfold=function(string,var="W",mode=-1) {
     }
     result
 }
-
-