+++ R/extractIMM.R
@@ -57,4 +57,3 @@ extractIMM=function(string){
     if(str_detect(result,"W")) return(NULL)
     result
 }
-