+++ R/tripleInteraction.R
@@ -12,10 +12,9 @@
 #' tripleInteraction(vars,mode=1)
 tripleInteraction=function(vars,prefix="c",suffix=0,mode=0,addPrefix=TRUE){
 
-    if(mode){
+    if (mode) {
         temp=paste(vars,collapse="*")
-
-    } else{
+    } else {
     result=vars
     result=c(result,apply(combn(vars,2),2,function(x){paste(x,collapse=":")}))
     result=c(result,paste0("interaction",suffix))
@@ -300,11 +299,11 @@ treatModerator=function(ind,moderatorNames,data=NULL,rangemode=1,probs=c(0.16,0.
 #'temp1= c("a1*hp","a2*wt","a3*hp:wt")
 #'temp2= "b1*am"
 #'temp3= c("c1*hp","c2*wt","c3*hp:wt")
-#'moderatorNames=c("wt")
-#'X= c("d1","d2");M="am"
-#'temp1=c("a1*d1","a2*d2","a3*wt","a4*d1:wt","a5*d2:wt")
-#'temp2="b1*am"
-#'temp3=c("c1*d1","c2*d2","c3*wt","c4*d1:wt","c5*d2:wt")
+#'#moderatorNames=c("wt")
+#'#X= c("d1","d2");M="am"
+#'#temp1=c("a1*d1","a2*d2","a3*wt","a4*d1:wt","a5*d2:wt")
+#'#temp2="b1*am"
+#'#temp3=c("c1*d1","c2*d2","c3*wt","c4*d1:wt","c5*d2:wt")
 #'cat(makeIndirectEquation(X,M,temp1,temp2,temp3,moderatorNames))
 #'cat(makeIndirectEquation(X,M,temp1,temp2,temp3,moderatorNames,range=TRUE))
 #'X="wt";M=NULL;temp1=NULL;temp2=NULL;temp3=c("c1*wt","c2*hp","c3*wt:hp");
@@ -423,4 +422,3 @@ makeIndirectEquation=function(X,M,temp1,temp2,temp3,moderatorNames,
 
     equation
 }
-