+++ NEWS.md
@@ -1,8 +1,17 @@
+## R package processR version 0.2.3
+===================================
+(2020-Jan-25)
+
+- Bug fixed
+
+- CRAN release
+
+
 ## R package processR version 0.2.2.6
 =====================================
 (2019-Nov-18)
 
-- independent on the package bda
+- independent on the package bda, jtools
 
 
 ## R package processR version 0.2.2.5